requirements.txt
appnope==0.1.0
decorator==4.0.2
Django==1.8.4
django-filter==0.11.0
djangorestframework==3.2.3
gnureadline==6.3.3
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
Markdown==2.6.2
path.py==8.1
pexpect==3.3
pickleshare==0.5
simplegeneric==0.8.1
traitlets==4.0.0
